Perform CLI tests simultaneously with python API
This should hopefully smooth out the development process by
preventing surprise CLI tests failures in CI/CD.
